+++ dsp/NoiseGate.h
@@ -32,7 +32,7 @@ const double MINIMUM_LOUDNESS_POWER = pow(10.0, MINIMUM_LOUDNESS_DB / 10.0);
 // The class that applies the gain reductions calculated by a trigger instance.
 class Gain : public DSP {
 public:
-  float **Process(float **inputs, const size_t numChannels,
+  double **Process(double **inputs, const size_t numChannels,
                   const size_t numFrames) override;
 
   void SetGainReductionDB(std::vector<std::vector<double>> &gainReductionDB) {
@@ -81,7 +81,7 @@ class Trigger : public DSP {
 public:
   Trigger();
 
-  float **Process(float **inputs, const size_t numChannels,
+  double **Process(double **inputs, const size_t numChannels,
                   const size_t numFrames) override;
   std::vector<std::vector<double>> GetGainReduction() const {
     return this->mGainReductionDB;
